bot/requirements.txt
# Telegram Forwarder Bot Dependencies

# Telegram client library
telethon==1.34.0

# Cryptography acceleration (optional but recommended)
cryptg==0.4.0

# PostgreSQL database driver (async)
asyncpg==0.29.0

# PostgreSQL database driver (sync - for migrations)
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# HTTP client (Telethon dependency)
aiohttp==3.9.1
